applications/requirements.txt
graphsignal==0.13.44
matplotlib==3.8.2 
numpy==1.26.2  
pandas==2.1.3
paramiko==3.3.1
plotly==5.18.0
requests==2.31.0
seaborn==0.13.0 
streamlit==1.29.0
altair==5.2.0                    
python-dotenv==1.0.0
langchain
openai
numexpr
chromadb
tiktoken
anthropic
